add testnet deposit toggle
