+++ src/views/dialogs/DepositDialog.tsx
@@ -1,23 +1,15 @@
-import { useState } from 'react';
-import styled, { type AnyStyledComponent } from 'styled-components';
-
-import { isMainnet } from '@/constants/networks';
 import { STRING_KEYS } from '@/constants/localization';
 import { useBreakpoints, useStringGetter } from '@/hooks';
 
 import { Dialog, DialogPlacement } from '@/components/Dialog';
 
-import { DepositForm } from '@/views/forms/AccountManagementForms/DepositForm';
-import { TestnetDepositForm } from '@/views/forms/AccountManagementForms/TestnetDepositForm';
-
-import { layoutMixins } from '@/styles/layoutMixins';
+import { DepositDialogContent } from './DepositDialog/DepositDialogContent';
 
 type ElementProps = {
   setIsOpen?: (open: boolean) => void;
 };
 
 export const DepositDialog = ({ setIsOpen }: ElementProps) => {
-  const [showFaucet, setShowFaucet] = useState(!isMainnet);
   const stringGetter = useStringGetter();
   const { isMobile } = useBreakpoints();
 
@@ -28,45 +20,9 @@ export const DepositDialog = ({ setIsOpen }: ElementProps) => {
       isOpen
       setIsOpen={setIsOpen}
       title={stringGetter({ key: STRING_KEYS.DEPOSIT })}
-      description={showFaucet && 'Test funds will be sent directly to your dYdX account.'}
       placement={isMobile ? DialogPlacement.FullScreen : DialogPlacement.Default}
     >
-      <Styled.Content>
-        {isMainnet || !showFaucet ? (
-          <DepositForm />
-        ) : (
-          <TestnetDepositForm onDeposit={closeDialog} />
-        )}
-        {!isMainnet && (
-          <Styled.TextToggle onClick={() => setShowFaucet(!showFaucet)}>
-            {showFaucet ? 'Show deposit form' : 'Show test faucet'}
-          </Styled.TextToggle>
-        )}
-      </Styled.Content>
+      <DepositDialogContent onDeposit={closeDialog} />
     </Dialog>
   );
 };
-
-const Styled: Record<string, AnyStyledComponent> = {};
-
-Styled.TextToggle = styled.div`
-  ${layoutMixins.stickyFooter}
-  color: var(--color-accent);
-  cursor: pointer;
-
-  margin-top: auto;
-
-  &:hover {
-    text-decoration: underline;
-  }
-`;
-
-Styled.Content = styled.div`
-  ${layoutMixins.stickyArea0}
-  --stickyArea0-bottomHeight: 2rem;
-  --stickyArea0-bottomGap: 1rem;
-  --stickyArea0-totalInsetBottom: 0.5rem;
-
-  ${layoutMixins.flexColumn}
-  gap: 1rem;
-`;
